requirements.txt
phonenumbers
requests
opencage
googlesearch-python
colorama
pycountry
geopy